fix bug2

+++ dupfilecheck.py
@@ -47,10 +47,14 @@ for (path, dir, files) in os.walk(sergdir):#지정한 파일 경로 탐색
             k.write(remv+"가"+max(pt,name[md.index(md5)])+"와 같은 파일인거 같아 삭제하였습니다. md5="+md5+"\n")#파일이 겹치면 해당 사실을 log.txt에 기록
             label1.config(text=remv+"가"+max(pt,name[md.index(md5)])+"와 같은 파일 인것 같아 삭제하였습니다.")
             window.update()
+            if(pt==name[md.index(md5)]):
+                fin()
+                continue
             try:
                 os.remove(remv)#겹치는 파일 삭제
             except:
                 os.remove(pt)
+                name[md.index(md5)]=pt
             if(min(remv)==pt):
                 name[md.index(md5)]=pt
         fin()
